Handle all nested types in yaml lists
For nested maps in a yaml list, rather than basing the map's contents on
the first list element, collect the contents of all of the list
elements. Then taking the distinct-named map elements, generate the
nested map's definition using the full list of data.

This means that the `database` field is now correctly processed from a
sample like this:

DbConnections:
    - Name: Primary
      Server: primary.host.net
    - Name: Dev
      Server: dev.somehost.net
      Database: dev1_db